+++ CHANGELOG.md
@@ -6,11 +6,29 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] v0.10.0
 
-## 9 Apr 2021
+## 10 Apr 2021
 
 ### Added
 
 ### Changed
+
+### Fixed
+
+### Removed
+
+## Released [v0.9.1]
+
+## 9 Apr 2021
+
+### Added
+web:
+- buttons demo
+
+### Changed
+slice-kit:
+- theme dark background & regular-text colors
+- button shadows with dark theme is black now
+- `base-button` and `stroked-button` has background by hover
 
 ### Fixed